fix: rename attestation start handler to attestation request handler in message and paired handlers
