+++ src/walletHandlers/attestationRequestHandler.js
